Fix style config
