+++ .gitignore
@@ -1,5 +1,5 @@
 .idea
 .git
-
+target/
