+++ cs_suppressions.xml
@@ -5,24 +5,5 @@
         "https://checkstyle.org/dtds/suppressions_1_2.dtd">
 
 <suppressions>
-    <suppress checks="MissingJavadocMethod"/>
-    <suppress checks="JavadocVariable"/>
-    <suppress checks="JavadocPackage"/>
-    <suppress checks="HiddenField"/>
-    <suppress checks="DesignForExtension"/>
-    <suppress checks="ParameterNumber"/>
-    <suppress checks="TypeName" files="main.java" lines="3"/>
-    <suppress checks="WhitespaceAround" files="Base.java" lines="19"/>
-    <suppress checks="WhitespaceAround" files="Chef.java" lines="14"/>
-    <suppress checks="WhitespaceAround" files="Customer.java" lines="14"/>
-    <suppress checks="WhitespaceAround" files="Ingredients.java" lines="18"/>
-    <suppress checks="WhitespaceAround" files="IoCContainer.java" lines="23"/>
-    <suppress checks="WhitespaceAround" files="Order.java" lines="17"/>
-    <suppress checks="WhitespaceAround" files="OrderDetail.java" lines="20"/>
-    <suppress checks="WhitespaceAround" files="Person.java" lines="16"/>
-    <suppress checks="WhitespaceAround" files="Product.java" lines="19"/>
-    <suppress checks="WhitespaceAround" files="main.java" lines="3"/>
-    <suppress checks="LineLength" files="Chef.java" lines="26"/>
-    <suppress checks="LineLength" files="Customer.java" lines="26"/>
-    <suppress checks="LineLength" files="IoCContainer.java" lines="35, 37, 44, 46"/>
+    <suppress checks="Javadoc" files="."/>
 </suppressions>
