+++ pre-commit
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+mvn clean
+mvn spotless:apply
+mvn checkstyle:check
