+++ src/main/java/com/ot/restaurant/Main.java
@@ -1,0 +1,3 @@
+package com.ot.restaurant;
+
+public class Main {}
