+++ src/main/java/com/ot/restaurant/constants/MeasureUnits.java
@@ -1,0 +1,24 @@
+package com.ot.restaurant.constants;
+
+public enum MeasureUnits {
+  TABLESPOONS("tbsp"),
+  TEASPOONS("tsp"),
+  CUPS("cup"),
+  GRAMS("g"),
+  MILLILITERS("ml"),
+  OUNCES("oz"),
+  POUNDS("lb"),
+  PINTS("pt"),
+  QUARTS("qt"),
+  LITERS("l");
+
+  private final String abbreviation;
+
+  MeasureUnits(final String abbreviation) {
+    this.abbreviation = abbreviation;
+  }
+
+  public String getAbbreviation() {
+    return abbreviation;
+  }
+}
