+++ src/main/java/com/ot/restaurant/constants/Status.java
@@ -1,0 +1,9 @@
+package com.ot.restaurant.constants;
+
+public enum Status {
+  ORDERED,
+  READY,
+  DELIVERED,
+  ACTIVE,
+  DELETED
+}
