+++ src/main/java/com/ot/restaurant/exceptions/DependencyKeyNullpointerException.java
@@ -1,0 +1,7 @@
+package com.ot.restaurant.exceptions;
+
+public class DependencyKeyNullpointerException extends Exception {
+  public DependencyKeyNullpointerException() {
+    super("Key Cannot be null");
+  }
+}
