+++ src/main/java/com/ot/restaurant/exceptions/DependencyNotFoundException.java
@@ -1,0 +1,7 @@
+package com.ot.restaurant.exceptions;
+
+public class DependencyNotFoundException extends Exception {
+  public DependencyNotFoundException() {
+    super("Dependency Value does not exists");
+  }
+}
