+++ src/main/java/com/ot/restaurant/exceptions/DependencyValueNullpointerException.java
@@ -1,0 +1,7 @@
+package com.ot.restaurant.exceptions;
+
+public class DependencyValueNullpointerException extends Exception {
+  public DependencyValueNullpointerException() {
+    super("Value cannot be null");
+  }
+}
