+++ src/main/java/com/ot/restaurant/exceptions/DuplicatedDependencyFoundException.java
@@ -1,0 +1,8 @@
+package com.ot.restaurant.exceptions;
+
+// Custom exception
+public class DuplicatedDependencyFoundException extends Exception {
+  public DuplicatedDependencyFoundException() {
+    super("Dependency already exists");
+  }
+}
