+++ src/main/java/com/ot/restaurant/images/3.3 customertEntity_UML.png
