--- src/main/java/com/restaurante/constants/Status.java
@@ -1,9 +1,0 @@
-package com.restaurante.constants;
-
-public enum Status {
-    ORDERED,
-    READY,
-    DELIVERED,
-    ACTIVE,
-    DELETED
-}
