--- src/main/java/com/restaurante/exceptions/DependencyKeyNullpointerException.java
@@ -1,7 +1,0 @@
-package com.restaurante.exceptions;
-
-public class DependencyKeyNullpointerException extends Exception {
-    public DependencyKeyNullpointerException() {
-        super("Key Cannot be null");
-    }
-}
