--- src/main/java/com/restaurante/exceptions/DependencyNotFoundException.java
@@ -1,7 +1,0 @@
-package com.restaurante.exceptions;
-
-public class DependencyNotFoundException extends Exception {
-    public DependencyNotFoundException() {
-        super("Dependency Value does not exists");
-    }
-}
