--- src/main/java/com/restaurante/exceptions/DependencyValueNullpointerException.java
@@ -1,7 +1,0 @@
-package com.restaurante.exceptions;
-
-public class DependencyValueNullpointerException extends Exception {
-    public DependencyValueNullpointerException() {
-        super("Value cannot be null");
-    }
-}
