--- src/main/java/com/restaurante/exceptions/DuplicatedDependencyFoundException.java
@@ -1,8 +1,0 @@
-package com.restaurante.exceptions;
-
-// Custom exception
-public class DuplicatedDependencyFoundException extends Exception {
-    public DuplicatedDependencyFoundException() {
-        super("Dependency already exists");
-    }
-}
