--- src/main/java/com/restaurante/main.java
@@ -1,3 +1,0 @@
-package com.restaurante;
-
-public class main {}
